[IMP] sale_coupon_promotion_generate_coupon, coupon_promotion_generate_coupon: split modules
TT42512
