+++ coupon_promotion_generate_coupon/README.rst
@@ -1,6 +1,6 @@
-===================================================
-Generate coupons in another coupon program in sales
-===================================================
+==========================================
+Generate coupons in another coupon program
+==========================================
 
 .. !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
    !! This file is generated by oca-gen-addon-readme !!
@@ -14,10 +14,10 @@ Generate coupons in another coupon program in sales
     :target: http://www.gnu.org/licenses/agpl-3.0-standalone.html
     :alt: License: AGPL-3
 .. |badge3| image:: https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github
-    :target: https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_promotion_generate_coupon
+    :target: https://github.com/OCA/sale-promotion/tree/15.0/coupon_promotion_generate_coupon
     :alt: OCA/sale-promotion
 .. |badge4| image:: https://img.shields.io/badge/weblate-Translate%20me-F47D42.png
-    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_promotion_generate_coupon
+    :target: https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_promotion_generate_coupon
     :alt: Translate me on Weblate
 .. |badge5| image:: https://img.shields.io/badge/runboat-Try%20me-875A7B.png
     :target: https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&target_branch=15.0
@@ -38,7 +38,9 @@ Usage
 
 To use this module, you need to have configured:
 
-#. A coupon program and a promotion program as told in ``coupon_promotion_generate_coupon`` readme.
+#. A coupon program which will be the one where the coupons will be generated.
+#. A promotion program with applicability on the next order and the first program
+   selected as next order program.
 
 Then, when a user meets the promotion program requisites, a pending coupon will be
 generated in the coupon program.
@@ -52,7 +54,7 @@ Bug Tracker
 Bugs are tracked on `GitHub Issues <https://github.com/OCA/sale-promotion/issues>`_.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_promotion_generate_coupon%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
+`feedback <https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_promotion_generate_coupon%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**>`_.
 
 Do not contact contributors directly about support or help with technical issues.
 
@@ -85,6 +87,6 @@ OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.
 
-This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_promotion_generate_coupon>`_ project on GitHub.
+This module is part of the `OCA/sale-promotion <https://github.com/OCA/sale-promotion/tree/15.0/coupon_promotion_generate_coupon>`_ project on GitHub.
 
 You are welcome to contribute. To learn how please visit https://odoo-community.org/page/Contribute.
