+++ coupon_promotion_generate_coupon/__init__.py
@@ -1,0 +1,1 @@
+from . import models
