+++ coupon_promotion_generate_coupon/readme/CONTRIBUTORS.rst
@@ -1,0 +1,4 @@
+* `Tecnativa <https://www.tecnativa.com>`_:
+
+  * David Vidal
+  * Pilar Vargas
