+++ coupon_promotion_generate_coupon/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module allows to generate pending coupons from a promotion to another one so the
+rules to apply the coupon could be different from the ones to generate it.
