+++ coupon_promotion_generate_coupon/readme/USAGE.rst
@@ -1,6 +1,8 @@
 To use this module, you need to have configured:
 
-#. A coupon program and a promotion program as told in ``coupon_promotion_generate_coupon`` readme.
+#. A coupon program which will be the one where the coupons will be generated.
+#. A promotion program with applicability on the next order and the first program
+   selected as next order program.
 
 Then, when a user meets the promotion program requisites, a pending coupon will be
 generated in the coupon program.
