+++ coupon_promotion_generate_coupon/static/description/index.html
@@ -4,7 +4,7 @@
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
 <meta name="generator" content="Docutils: http://docutils.sourceforge.net/" />
-<title>Generate coupons in another coupon program in sales</title>
+<title>Generate coupons in another coupon program</title>
 <style type="text/css">
 
 /*
@@ -360,14 +360,14 @@ ul.auto-toc {
 </style>
 </head>
 <body>
-<div class="document" id="generate-coupons-in-another-coupon-program-in-sales">
-<h1 class="title">Generate coupons in another coupon program in sales</h1>
+<div class="document" id="generate-coupons-in-another-coupon-program">
+<h1 class="title">Generate coupons in another coupon program</h1>
 
 <!-- !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 !! This file is generated by oca-gen-addon-readme !!
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
-<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_promotion_generate_coupon"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-sale_coupon_promotion_generate_coupon"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&amp;target_branch=15.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
+<p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/coupon_promotion_generate_coupon"><img alt="OCA/sale-promotion" src="https://img.shields.io/badge/github-OCA%2Fsale--promotion-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/sale-promotion-15-0/sale-promotion-15-0-coupon_promotion_generate_coupon"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runboat.odoo-community.org/webui/builds.html?repo=OCA/sale-promotion&amp;target_branch=15.0"><img alt="Try me on Runboat" src="https://img.shields.io/badge/runboat-Try%20me-875A7B.png" /></a></p>
 <p>This module allows to generate pending coupons from a promotion to another one so the
 rules to apply the coupon could be different from the ones to generate it.</p>
 <p><strong>Table of contents</strong></p>
@@ -387,7 +387,9 @@ rules to apply the coupon could be different from the ones to generate it.</p>
 <h1><a class="toc-backref" href="#id1">Usage</a></h1>
 <p>To use this module, you need to have configured:</p>
 <ol class="arabic simple">
-<li>A coupon program and a promotion program as told in <tt class="docutils literal">coupon_promotion_generate_coupon</tt> readme.</li>
+<li>A coupon program which will be the one where the coupons will be generated.</li>
+<li>A promotion program with applicability on the next order and the first program
+selected as next order program.</li>
 </ol>
 <p>Then, when a user meets the promotion program requisites, a pending coupon will be
 generated in the coupon program.</p>
@@ -399,7 +401,7 @@ configured in that promotion.</p>
 <p>Bugs are tracked on <a class="reference external" href="https://github.com/OCA/sale-promotion/issues">GitHub Issues</a>.
 In case of trouble, please check there if your issue has already been reported.
 If you spotted it first, help us smashing it by providing a detailed and welcomed
-<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20sale_coupon_promotion_generate_coupon%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
+<a class="reference external" href="https://github.com/OCA/sale-promotion/issues/new?body=module:%20coupon_promotion_generate_coupon%0Aversion:%2015.0%0A%0A**Steps%20to%20reproduce**%0A-%20...%0A%0A**Current%20behavior**%0A%0A**Expected%20behavior**">feedback</a>.</p>
 <p>Do not contact contributors directly about support or help with technical issues.</p>
 </div>
 <div class="section" id="credits">
@@ -427,7 +429,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <p>OCA, or the Odoo Community Association, is a nonprofit organization whose
 mission is to support the collaborative development of Odoo features and
 promote its widespread use.</p>
-<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/sale_coupon_promotion_generate_coupon">OCA/sale-promotion</a> project on GitHub.</p>
+<p>This module is part of the <a class="reference external" href="https://github.com/OCA/sale-promotion/tree/15.0/coupon_promotion_generate_coupon">OCA/sale-promotion</a> project on GitHub.</p>
 <p>You are welcome to contribute. To learn how please visit <a class="reference external" href="https://odoo-community.org/page/Contribute">https://odoo-community.org/page/Contribute</a>.</p>
 </div>
 </div>
