+++ coupon_promotion_generate_coupon/tests/__init__.py
@@ -1,0 +1,1 @@
+from . import test_coupon_promotion_generate_coupon
