+++ setup/coupon_promotion_generate_coupon/odoo/addons/coupon_promotion_generate_coupon
@@ -1,0 +1,1 @@
+../../../../coupon_promotion_generate_coupon
